backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
httpx
firebase-admin
